LL-6768 - CI - fix comment when all images change (#4086)
* test

* try to change all screenshots

* try again

* change upload script

* merge comment step into one and read from file directly

* fixes

* remove screenshot breaker

* naming
