+++ .github/actions/comment/action.yml
@@ -1,11 +1,7 @@
-name: prepare-comment-body
+name: comment
 description: gets a list of images and return a formatted markdown body for comment
 author: valpinkman
 outputs:
-  body:
-    description: json formatted body for comment
-  bodyclean:
-    description: markdown formatted body for comment
   bodySlack:
     description: message formatted body for slack
   bodySlackAuthor:
@@ -13,8 +9,11 @@ outputs:
   slackAuthor:
     description: slack handle of the author
 inputs:
+  prNumber:
+    description: prNumber
+    required: true
   images:
-    description: array of images to include in body of comment
+    description: stringified array of images to include in body of comment
     required: true
   fullrepo:
     description: full repo name
