--- .github/actions/post-comment/action.yml
@@ -1,17 +1,0 @@
-name:  post-comment
-description: post comment
-author: valpinkman
-inputs:
-  comment:
-    description: comment
-    required: true
-  prNumber:
-    description: prNumber
-    required: true
-runs:
-  using: node12
-  main: index.js
-
-branding:
-  icon: align-center
-  color: white
