--- .github/actions/post-comment/index.js
@@ -1,20 +1,0 @@
-const core = require("@actions/core");
-const fetch = require("isomorphic-unfetch");
-
-const main = async () => {
-  const comment = core.getInput("comment");
-  const prNumber = core.getInput("prNumber");
-
-  await fetch(
-    `http://github-actions-live-vercel.vercel.app/api/comment?owner=LedgerHQ&repo=ledger-live-desktop&issueId=${prNumber}`,
-    {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: comment,
-    },
-  );
-};
-
-main().catch(err => core.setFailed(err));
