+++ .github/workflows/checks.yml
@@ -47,12 +47,6 @@ jobs:
       - name: run code checks
         if: always() && job.status != 'cancelled'
         run: yarn ci 2> >(tee lint.txt)
-      - name: Read lint output
-        id: lint
-        if: always() && job.status != 'cancelled'
-        uses: juliangruber/read-file-action@e0a316da496006ffd19142f0fd594a1783f3b512
-        with:
-          path: ./lint.txt
       - name: check build
         if: always() && job.status != 'cancelled'
         env:
@@ -61,12 +55,6 @@ jobs:
       - name: Tests
         if: always() && job.status != 'cancelled'
         run: yarn test 2> >(tee jest.txt)
-      - name: Read jest output
-        id: jest
-        if: always() && job.status != 'cancelled'
-        uses: juliangruber/read-file-action@e0a316da496006ffd19142f0fd594a1783f3b512
-        with:
-          path: ./jest.txt
       - name: start electron webdriver
         if: always() && job.status != 'cancelled'
         env:
@@ -80,12 +68,6 @@ jobs:
       - name: run spectron
         if: always() && job.status != 'cancelled'
         run: yarn spectron --no-color 2> >(tee output.txt)
-      - name: Read test output
-        id: spectron
-        if: always() && job.status != 'cancelled'
-        uses: juliangruber/read-file-action@e0a316da496006ffd19142f0fd594a1783f3b512
-        with:
-          path: ./output.txt
       - name: run coverage generation
         if: always() && github.event_name == 'push' && job.status != 'cancelled'
         run: yarn spectron-coverage
@@ -112,40 +94,22 @@ jobs:
           git fetch ledger
           git diff --name-only ledger/${{github.base_ref}}..${{github.event.pull_request.head.sha}} -- tests/specs/__image_snapshots__ >imgChanged.txt
         if: always() && github.event_name == 'pull_request'
-      - name: Read imgChanged output
-        id: imgChanged
-        if: always() && github.event_name == 'pull_request' && job.status != 'cancelled'
-        uses: juliangruber/read-file-action@e0a316da496006ffd19142f0fd594a1783f3b512
-        with:
-          path: ./imgChanged.txt
-      - name: prepare comment body
+      - name: send comment
         if: always() && job.status != 'cancelled'
         id: comment
-        uses: ./.github/actions/prepare-comment-body
+        uses: ./.github/actions/comment
         with:
           images: ${{ steps.imgur.outputs.images }}
           fullrepo: ${{ github.repository }}
           runId: ${{ github.run_id }}
-          testoutput: ${{ steps.spectron.outputs.content }}
-          lintoutput: ${{ steps.lint.outputs.content }}
-          jestoutput: ${{ steps.jest.outputs.content }}
+          testoutput: ./output.txt
+          lintoutput: ./lint.txt
+          jestoutput: ./jest.txt
+          imgChanged: ./imgChanged.txt
           author: ${{github.event.sender.login}}
           pullId: ${{github.event.number}}
-          imgChanged: ${{steps.imgChanged.outputs.content}}
           from: ${{github.base_ref}}
           to: ${{github.event.pull_request.head.sha}}
-      - name: display comment
-        if: always() && job.status != 'cancelled'
-        id: displaycomment
-        env:
-          BODY: ${{ steps.comment.outputs.bodyclean }}
-        run: echo "$BODY"
-      - name: post comment
-        if: always() && job.status != 'cancelled'
-        id: postcomment
-        uses: ./.github/actions/post-comment
-        with:
-          comment: ${{ steps.comment.outputs.body }}
           prNumber: ${{github.event.pull_request.number}}
       - name: Send message to Slack channel
         uses: archive/github-actions-slack@v1.0.3
